Add bl_info for work as plugin
Adjust FoamBlocks to its complete block size and no 1/2 size

Adjust the gap between blocks in the Z-axis to 0.002 meters.

Rotate the Y-wood 90 degrees in the X-axis.

Fix the error that occurs when attempting to create an STL file when no
folder path exists.

When exporting to STL, rotate it -90 or 90 degrees in the Y-axis,
depending on whether the object is to the right or left relative to the
center of the original irregularly cut objectrepect to the middle of the
original cuted irregular object