Made the "context" available  for the functions when needed, this change the global variables by context, global variables don't work when the script is installed.
When it cut the iregular objec, it calculate the total volume of the part that has just been cut, and also counts how many parts there are.

All the files will be created in a directory called "PARTES".

At the moment of export the g_code_raw files, it first delete all the previus g_code_raw files.

Modified  the logic of rotate the cut part related to the middle of the complete-irregular-object, making the rotation only for the first layer in Z axis.

Fix logic problems in the get_total_area function and changed it to get_total_area_vol adding logic to calculate volume.